arm build fixes
This is a monster bunch.  Firstly, eliminate the efi call wrapper
thunking.  On the security policy override, this was done via an
x86_64 asm routine which won't work on non-x86.

The build arm objects using the -O binary objcopy method and take the
linker scripts from gnu-efi rather than hand rolling.

Confine the EFI building machinery to its own include file (not having
this correct was causing an OBS build failure on arm)

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>