shim_protocol: add protocol installer
The shim_protocol is necessary because it is used as a callback by
grub to verify the signature of linux kernel images.  This means that
even if you're relying entirely on the secure boot keys, you cannot
simply replace shim.efi with grub.efi.  Add a shim protocol installer
in preparation for adding a replacement shim that does nothing other
than install the protocols and execute grub.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>