ShimReplace: add new shim loader simply to install protocol
The way grub currently works on Linux is that it relies on the shim
protocol to verify images.  Without this, the secure boot chain is
broken.  Fix this by adding a shim replacement whose sole job is to
install the protocol and call the boot loader via the normal fashion
(meaning the bootloader must be signed with a key in the secure boot
database).  The second stage loader can then use the protocol to
verify any images against the secure boot database as well.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>