cert-to-efi-sig-list: allow multiple certs in the sig list
This patch adds the ability to bundle multiple certificates in an
EFI_SIGNATURE_LIST.  It makes no attempt to optimize the list for size,
because for certs that's not likely to succeed at all, and hashes are
still not supported.

The syntax remains as it was, with the extra certs being added before
the final command line argument (i.e. before the output) as in:

cert-to-efi-sig-list -g $GUID DB.crt DB1.crt DB2.crt DB.esl

Signed-off-by: Peter Jones <pjones@redhat.com>