Work around some unfortunate fs path choices pjones made on RHEL and Fedora.
On Fedora and RHEL, libgnuefi.a is in /usr/lib{,64}/ and crt0.o and
elf.lds are both in /usr/lib{,64}/gnuefi/.

This tries to find each of those in a similar method to the existing
CRTPATH test loop.

We also have the include directories having the EFI arch name now.

Signed-off-by: Peter Jones <pjones@redhat.com>